🐛 fixes "too many items" when names are null  (#543)
* 🐛 avert infinite loop when prod_id not assigned

Without ptr being reassigned in the case of a 0 prod_id, the app would get into an infinite loop on launch, before opening- this should resolve it 😊

* 🐛 fixes "too many items" when names are null

There is an (honestly misleading) OutOfMemoryException with the message "too many items in listbox" that is thrown when there are any items whose ToString() method returns null. This commit (in my maybe-biased opinion) gives a reasonable output in that case instead.