Edited for readability and consistency
The comments in the original file are difficult to read with a text editor that is not colour coded due to the amount of single line comments. Users with no coding knowledge will not understand what they are looking at. The layout of comments and settings were also inconsistent from setting to setting. The comments have been updated to more adequately explain what each setting does and collated into a single block comment. The order of settings has also been changed to more logically group related settings. Lastly, a comment at the top of the file explaining how to edit the settings has been added for inexperienced users, with a suggestion to use an editor other than Windows Notepad.