Edited readme to reflect state of wiki
The wiki is woefully inadequate, and given the amount of users creating issues that end up being closed with links to the wiki and "RTFM," it's abundantly clear the issue is the lack of proper documentation. The link to the wiki now has clarification that the wiki is still under construction to reflect this.